+++ pytorch_lightning/trainer/configuration_validator.py
@@ -357,12 +357,9 @@ def _check_deprecated_callback_hooks(trainer: "pl.Trainer") -> None:
                     f"The `Callback.{hook}` hook was deprecated in v1.6 and"
                     f" will be removed in v1.8. Please use `Callback.{alternative_hook}` instead."
                 )
-        for hook, alternative_hook in (
-            ("on_pretrain_routine_start", "on_fit_start"),
-            ("on_pretrain_routine_end", "on_fit_start"),
-        ):
+        for hook in ("on_pretrain_routine_start", "on_pretrain_routine_end"):
             if is_overridden(method_name=hook, instance=callback):
                 rank_zero_deprecation(
-                    f"The `Callback.{hook}` hook was deprecated in v1.6 and"
-                    f" will be removed in v1.8. Please use `Callback.{alternative_hook}` instead."
+                    f"The `Callback.{hook}` hook has been deprecated in v1.6 and"
+                    f" will be removed in v1.8. Please use `Callback.on_fit_start` instead."
                 )