+++ tests/deprecated_api/test_remove_1-8.py
@@ -210,6 +210,8 @@ def test_v1_8_0_deprecate_trainer_callback_hook_mixin():
         "on_epoch_end",
         "on_train_start",
         "on_train_end",
+        "on_pretrain_routine_start",
+        "on_pretrain_routine_end",
         "on_batch_start",
         "on_batch_end",
         "on_validation_start",
@@ -245,7 +247,8 @@ def test_v1_8_0_deprecate_trainer_callback_hook_mixin():
         logger=False,
     )
     model = BoringModel()
-    trainer.fit(model)
+    # need to attach model to trainer for testing of `on_pretrain_routine_start`
+    trainer.strategy.connect(model)
     for method_name in methods_with_self:
         fn = getattr(trainer, method_name, None)
         with pytest.deprecated_call(match="was deprecated in v1.6 and will be removed in v1.8"):